+++ src/services/postService.js
@@ -28,3 +28,26 @@ export async function fetchRecentPosts(page) {
         throw error;
     }
 }
+
+//질문이야기 게시판 조회 로직
+export async function questionShow() {
+    try {
+        const response = await fetch(
+            `${process.env.NEXT_PUBLIC_BASE_URL}/board/question/show`,
+            {
+                method: 'GET',
+            },
+        );
+
+        if (!response.ok) {
+            throw new Error(`Error: ${response.status}`);
+        }
+
+        const data = await response.json();
+        console.log(data);
+        return data;
+    } catch (error) {
+        console.error('Error fetching data', error);
+        return [];
+    }
+}
